Rework SkipNode and add insertion.
==== On SkipNodes ====

All Option<*mut SkipNode<T>> are replaced with *mut SkipNode<T>.

Null pointer checking are replaced with ptr::to_{ref/mut}() whenever
possible.

When you calculate distance between nodes, head used to be a special
case that does not contribute to distance, which IMHO is not that intuitive.

Now all distance related methods are moved to SkipNode.

=== On insertion =====

I split the old methods into smaller methods, and use recursion
instead of loops to handle different levels.

Now the implementation is generalized a bit by allowing the caller to
control where they want to insert the object, at the cost of making the
code more complicated.  Hopefully those methods can be shared among
different data structures, and unsafe regions can be minimized.
One may argue it's overengineering, and I don't disagree.
But I wrote it as a hobby, so it seems okay.

One side benefit of using recursion is that it's more obvious which
sections of code are for clean up / link fixing. And there's no
heap allocation. But it uses more stack space.

Add a new test for insertion.