Make all data structures covariant.
It is now possible to convert SkipList<&'static str> to SkipList<&'a str>
for any 'a.

In in formal terms, any type with longer life time can usually be
converted into "compitable" types with shorter lifetime.
e.g. You can store `&'static str` in `&'a str` for all `'a`,
because `&'a str` is "compitble" with `&'static str`.

However with data structures like `SkipList` it's more difficult for
the compiler to see it's okay to store `SkipList<'static str>` in a
`SkipList<'a str>`.

Covariance is a more formal way to say "If T is compitable to U, then
`SkipList<T>` is compitable to `SkipList<U>`".

Data structures in general should support this behavior. For example,
`Vec<T>` is covariant over T, e.g. you can convert `Vec<&'static str>`
to `Vec<&'a str>` for any 'a.  This patch makes all data strutures in
this crate covariant.

For details see Rustonomicon.
https://doc.rust-lang.org/nomicon/subtyping.html